+++ provers/cairo/Makefile
@@ -2,15 +2,37 @@
 
 ROOT_DIR:=$(shell dirname $(realpath $(firstword $(MAKEFILE_LIST))))
 
+CAIRO0_PROGRAMS_DIR=cairo_programs/cairo0
+CAIRO0_PROGRAMS:=$(wildcard $(CAIRO0_PROGRAMS_DIR)/*.cairo)
+COMPILED_CAIRO0_PROGRAMS:=$(patsubst $(CAIRO0_PROGRAMS_DIR)/%.cairo, $(CAIRO0_PROGRAMS_DIR)/%.json, $(CAIRO0_PROGRAMS))
+
 # Variable to give compiled Cairo programas a proper name. It extracts the file extension and adds
 # the .json extension to it.
-COMPILED_PROGRAM=$(basename $(PROGRAM_PATH)).json
+COMPILED_PROGRAM=$(basename $(PROGRAM)).json
+
+# Rule to compile Cairo programs for testing purposes.
+# If the `cairo-lang` toolchain is installed, programs will be compiled with it.
+# Otherwise, the cairo_compile docker image will be used
+# When using the docker version, be sure to build the image using `make docker_build_cairo_compiler`.
+$(CAIRO0_PROGRAMS_DIR)/%.json: $(CAIRO0_PROGRAMS_DIR)/%.cairo
+	@echo "Compiling Cairo program..."
+	@cairo-compile --cairo_path="$(CAIRO0_PROGRAMS_DIR)" $< --output $@ 2> /dev/null --proof_mode || \
+	docker run --rm -v $(ROOT_DIR)/$(CAIRO0_PROGRAMS_DIR):/pwd/$(CAIRO0_PROGRAMS_DIR) cairo --proof_mode /pwd/$< > $@
 
 build: 
 	cargo build --release
 
+test: $(COMPILED_CAIRO0_PROGRAMS)
+	cargo test
+
+test_metal: $(COMPILED_CAIRO0_PROGRAMS)
+	cargo test -F metal
+
 docker_build_cairo_compiler:
 	docker build -f cairo_compile.Dockerfile -t cairo .
+
+docker_compile_cairo:
+	docker run -v $(ROOT_DIR):/pwd cairo --proof_mode /pwd/$(PROGRAM) > $(COMPILED_PROGRAM)
 
 compile:
 	cairo-compile --proof_mode $(PROGRAM_PATH) > $(COMPILED_PROGRAM) 2>/dev/null || \
@@ -25,13 +47,6 @@ verify:
 run_all:
 	cargo run --quiet --release prove-and-verify $(PROGRAM_PATH)
 
-compile_and_prove:
-	@cairo-compile --proof_mode $(PROGRAM_PATH) > $(COMPILED_PROGRAM) 2>/dev/null || \
-	docker run -v $(ROOT_DIR):/pwd cairo --proof_mode /pwd/$(PROGRAM_PATH) > $(COMPILED_PROGRAM)
-	@echo "Compiling done \n"
-	@cargo run --quiet --release prove $(COMPILED_PROGRAM) $(PROOF_PATH)
-	@rm $(COMPILED_PROGRAM)
-
 compile_and_run_all:
 	@cairo-compile --proof_mode $(PROGRAM_PATH) > $(COMPILED_PROGRAM) 2>/dev/null || \
 	docker run -v $(ROOT_DIR):/pwd cairo --proof_mode /pwd/$(PROGRAM_PATH) > $(COMPILED_PROGRAM)
@@ -41,3 +56,46 @@ compile_and_run_all:
 
 clippy:
 	cargo clippy --workspace --all-targets -- -D warnings
+
+
+benchmarks_sequential: $(COMPILED_CAIRO0_PROGRAMS)
+	cargo bench
+
+benchmarks_parallel: $(COMPILED_CAIRO0_PROGRAMS)
+	cargo bench -F parallel --bench criterion_prover
+	cargo bench -F parallel --bench criterion_verifier
+
+benchmarks_parallel_all: $(COMPILED_CAIRO0_PROGRAMS)
+	cargo bench -F parallel
+
+build_metal:
+	cargo b --features metal --release
+
+target/release/cairo-platinum-prover:
+	cargo build --bin cairo-platinum-prover --release --features instruments
+	
+docker_compile_and_run_all: target/release/cairo-platinum-prover
+	@echo "Compiling program with docker"
+	docker run -v $(ROOT_DIR):/pwd cairo --proof_mode /pwd/$(PROGRAM) > $(COMPILED_PROGRAM)
+	@echo "Compiling done \n"
+	@cargo run --bin cairo-platinum-prover --features instruments --quiet --release prove_and_verify $(COMPILED_PROGRAM)
+	@rm $(COMPILED_PROGRAM)
+
+docker_compile_and_prove: target/release/cairo-platinum-prover
+	@echo "Compiling program with docker"
+	@docker run -v $(ROOT_DIR):/pwd cairo --proof_mode /pwd/$(PROGRAM) > $(COMPILED_PROGRAM)
+	@echo "Compiling done \n"
+	@cargo run --bin cairo-platinum-prover --features instruments --quiet --release prove $(COMPILED_PROGRAM) $(PROOF_PATH)
+	@rm $(COMPILED_PROGRAM)
+
+compile_and_prove: target/release/cairo-platinum-prover
+	@echo "Compiling program with cairo-compile"
+	@cairo-compile --proof_mode $(PROGRAM) --output $(COMPILED_PROGRAM)
+	@echo "Compiling done \n"
+	@cargo run --bin cairo-platinum-prover --features instruments --quiet --release prove $(COMPILED_PROGRAM) $(PROOF_PATH)
+	@rm $(COMPILED_PROGRAM)
+
+clean:
+	rm -f $(CAIRO0_PROGRAMS_DIR)/*.json
+	rm -f $(CAIRO0_PROGRAMS_DIR)/*.trace
+	rm -f $(CAIRO0_PROGRAMS_DIR)/*.memory
