+++ provers/cairo/benches/criterion_prover.rs
@@ -1,14 +1,14 @@
-use cairo_platinum_prover::cairo_layout::CairoLayout;
-use cairo_platinum_prover::{air::generate_cairo_proof, runner::run::generate_prover_args};
+use platinum_prover::{
+    air::generate_cairo_proof, cairo_layout::CairoLayout, runner::run::generate_prover_args,
+};
 use criterion::{
     black_box, criterion_group, criterion_main, measurement::WallTime, BenchmarkGroup, Criterion,
-    SamplingMode,
 };
 use stark_platinum_prover::proof::options::{ProofOptions, SecurityLevel};
 
 pub mod functions;
 
-fn fibo_70k_bench(c: &mut Criterion) {
+fn cairo_benches(c: &mut Criterion) {
     #[cfg(feature = "parallel")]
     {
         let num_threads: usize = std::env::var("NUM_THREADS")
@@ -23,12 +23,17 @@ fn fibo_70k_bench(c: &mut Criterion) {
     };
 
     let mut group = c.benchmark_group("CAIRO");
-    group.sampling_mode(SamplingMode::Flat);
     group.sample_size(10);
     run_cairo_bench(
         &mut group,
-        "fibonacci/70000",
-        &cairo0_program_path("fibonacci_70000.json"),
+        "fibonacci/500",
+        &cairo0_program_path("fibonacci_500.json"),
+        CairoLayout::Plain,
+    );
+    run_cairo_bench(
+        &mut group,
+        "fibonacci/1000",
+        &cairo0_program_path("fibonacci_1000.json"),
         CairoLayout::Plain,
     );
 }
@@ -49,6 +54,7 @@ fn run_cairo_bench(
     let program_content = std::fs::read(program_path).unwrap();
     let proof_options = ProofOptions::new_secure(SecurityLevel::Provable80Bits, 3);
     let (main_trace, pub_inputs) = generate_prover_args(&program_content, &None, layout).unwrap();
+    println!("Generated main trace with {} rows", main_trace.n_rows());
 
     group.bench_function(benchname, |bench| {
         bench.iter(|| {
@@ -57,5 +63,5 @@ fn run_cairo_bench(
     });
 }
 
-criterion_group!(benches, fibo_70k_bench);
+criterion_group!(benches, cairo_benches);
 criterion_main!(benches);
