+++ provers/cairo/cairo_programs/cairo0/compare_lesser_array.cairo
