+++ provers/cairo/cairo_programs/cairo0/lt_comparison.cairo
