+++ provers/cairo/src/errors.rs
