+++ provers/cairo/src/runner/vec_writer.rs
